Rewrite of busy-signal
Summary:
This is a rewrite of the busy-signal to serve some needs from
the language services:

(1) Need to have progress messages that update over time,
e.g. "initializing [1/50]", "initializing [3/50]", ...

(2) Need to have a way to indicate 'action required' (i.e. it's not
that the user is busy waiting for the computer, but rather the
computer is busy waiting for the user. That's why conceptually it goes
in the same busy-signal service and the same spot in the status
bar). And this action-required needs to be clickable to help the user
perform their action.

(3) Need debouncing, so that users of this package can fire the busy
signal for all their work without having to keep track of how long the
work is.

Some other notes:

* Debounce duration is left to the discretion of the busy-signal
  package. Currently 300ms. Consumers can't pick the duration.

* Previously, busy-signal took in HTML strings, e.g. "this
  <b>bold</b>" would be rendered in the tooltip with bold. No
  consumers ever used it. I switched over to just taking strings that
  will be displayed exactly as-is. Makes things easier.

* Currently, for the UI, I made it so if something has action
  required, and something else is busy, then the status bar only shows
  the icon for 'action-required'. This could easily be changed in
  future. But I couldn't figure out an icon that combined the two
  nicely. Let's see how it goes

* The old busy-signal had a de-duplication feature where if two
  consumers had the exact same message to display then it would only
  show once. I removed that because I think it would be actively
  confusing - imagine if it went from "progress 10% / progress 25%" to
  "progress 50%" to "progress 60% / progress 70%" - you'd be confused
  why one of the progress reports had disappeared! So I removed this feature.

* In the future we might wish to add progress-bars, by adding a new
  method 'setProgress' alongside 'setTitle'. I didn't go there for
  now.

Reviewed By: matthewwithanm

Differential Revision: D5431155

fbshipit-source-id: bfc833b1990898215dc530a14339bee2b499697b