Make the compilation db entries a bit stricter
Summary:
now that hansonw has fixed the # problem with shellParse, I'm making this struct have a string all the time.
I'm curious about why it was string | Array<string>. I looked a bit in the history of the file but found nothing useful.
The official documentation (https://clang.llvm.org/docs/JSONCompilationDatabase.html) says `command: The compile command executed. After JSON unescaping, this must be a valid command to rerun the exact compilation step for the translation unit in the environment the build system uses`.

I also renamed flags to command, because that's actually what we are expecting.

Reviewed By: hansonw

Differential Revision: D5406286

fbshipit-source-id: 05214ce866fe30b5e00fbc475c8667cf0ffb9b5a