Extend the clang database providers
Summary:
I want these providers to be more powerful in order to fully remove buck from clang-rpc. The last piece is to provide the projectRoot from the client.
In order to achieve this, I'm extending the clang database provider to clang request settings provider.
This new object contains the compilation database along with the project root.

Reviewed By: hansonw

Differential Revision: D5410550

fbshipit-source-id: eb7b8d8a39d61d3405b77e54e1cceae08ea70de5