Transition bound-in-contructor methods to bound arrow functions as property initializers
Summary:
I got tired of missing out on all the fun I was used to with this on www, and our babel infrastructure
through transform-class-properties already supports this.

This was done entirely with vjeux's
[manual-bind-to-arrow](https://github.com/reactjs/react-codemod/blob/master/transforms/manual-bind-to-arrow.js).

It doesn't catch every use, so I'm going to spend a bit of time finding out why and if it's something easily contributable
back to the codemod. Otherwise I follow up with a separately revertable diff.

Reviewed By: vjeux

Differential Revision: D5476160

fbshipit-source-id: 1b41d1b02be33e1ca10b05b4693d23d905391dda