Clang: Manually dedupe busy signal messages
Summary:
For Clang lints, it's difficult to cancel pending lint requests since they're offloaded to an off-process language server (that doesn't use LSP, as it predates that).

As busy signal no longer automatically de-dupes, we can replicate that manually.
I used a `SharedObservableCache` to hold observables that report busy signals and then dispose them on unsubscription.

Reviewed By: matthewwithanm

Differential Revision: D5465740

fbshipit-source-id: cb7373c5e75865878d171325f1e7eb9108277e27