Start without debugging callback
Summary: Not sure if this is just me, but the `await consumeFirstProvider` call seems to hang forever on my machine, meaning the observable never produces a platform group, and the devices spinner spins forever instead of changing to a dropdown. Let's start by producing `null` which fixes this and should still immediately enable the debug button if the service is in fact consumed. Hopefully I'll come up with a better followup diff.

Reviewed By: ebluestein

Differential Revision: D5482443

fbshipit-source-id: efd859f2a76c431a66dafb248f5c78298aa9c4e0