Make createMultiLspLanguageService truly RPC-able [4/n]
Summary:
Loggers aren't a good idea to put over the RPC. Instead, we can just pass in a logging category + level. Note that `log4js` stores a singleton logging object per category, so this is practically identical.

The list of options is way too long so I moved `languageId`, `command`, and `args` to the top, and moved the rest to a `params` object. Open to other renaming suggestions.

Reviewed By: arxanas

Differential Revision: D5475126

fbshipit-source-id: 6ccb7ff9f43d597cef09e4fd608c1fe2f87d91c4