Include ALT accelerator key for Nuclide menu on Windows
Summary:
On Windows, menus are expected to define an accelerator key by placing an & before a letter in the menu label. This & will not appear on screen, but causes ALT+N to open the Nuclide menu.

All of the built in Atom menus are providing this correctly, but the Nuclide menu is not.

Accelerator keys are handy for keyboard users, and are required for assistive technologies like screen readers to be able to navigate properly.

Reference: https://msdn.microsoft.com/en-us/library/windows/desktop/ms646335(v=vs.85).aspx#_win32_Accelerators_and_Menus

Reviewed By: hansonw

Differential Revision: D5491860

fbshipit-source-id: c55609bfcadb89a62be89a70c500cb5482f4bd0b