Support finding related headers
Summary:
Previously, the clang plugin only supported finding related sources. Now this adds the corresponding logic for related headers.

This is useful for the `swith to header/source` button.

For normal projects this was never an issue because the related file finder plugin was looking for the related headers in the same folder as the source, thus not needing to call the clang plugin.
However, for projects with headers in different locations, like in the case of obj-c frameworks, a special logic was required.

Reviewed By: sas

Differential Revision: D5463083

fbshipit-source-id: 7e6d9ac7f55d834525ce1e34310f5822b7848769