Persistent Connections Filter Warnings By File (Nuclide-Side)
Summary:
This is the Nuclide code of a change to the persistent connection protocol. The
Flow code is in D5439642.

Changed the persistent connection protocol to allow clients to track
and untrack files. Warnings are only shown on tracked files. This allows
Nuclide to track only open files, which keeps large numbers of existing
warnings from overwhelming the persistent connection.

Reviewed By: nmote

Differential Revision: D5439741

fbshipit-source-id: 9f255b545428cfb709b2971cca8da65792341bdd