AutoCompleteCacher avoids issuing additional lang-server requests in some cases
Summary:
The code for AutoCompleteCacher had this comment:

> We need to send this request speculatively because if firstResult resolves to `null`, we'll
> need this result. If we wait for firstResult to resolve before sending it, satisfying this
> request could take as much as two round trips to the server. We could avoid this in some
> cases by checking if firstResult has already been resolved. If it has already resolved to a
> non-null value, we can skip this request.

This diff makes the change proposed by that comment.

Motive: (1) I think it's confusing for a language-service author to see multiple requests coming their way when they believe it shouldn't be needed. It was for me. (2) I want to investigate why Hack has a 400ms latency for subsequent filtering of a 1000+ member class, and want to tighten up on the caching.

Reviewed By: nmote

Differential Revision: D5368993

fbshipit-source-id: de8750d46cb88f47783f91a7f7b2a0a90b9ed091