Make autocomplete use language-appropriate replacement prefix
Summary:
'prefix' has to do with what's replaced when the user accepts an
autocomplete suggestion. It's based on the current word. For instance,
  '$c|'      => suggestion '$compare'  => hopefully replace '$c'
  'Vec\com|' => suggestion 'compare'   => hopefully replace 'com'
The way autocomplete works is: the language service might say what prefix its suggestion will replace; and if it doesn't, then autocomplete will replace whatever prefix was part of the 'request' object.

Atom has its own way of computing the current word (to support gestures like cursor-past-word). It bases this on 'editor.nonWordCharacters', by default roughly `[a-zA-Z0-9_]`. But language packages can and do override this -- e.g. PHP allows '$' in identifiers.

Autocomplete doesn't use Atom's technique (I suspect because the html and css and xml packages never bothered overriding it). Instead autocomplete has its own regex, roughly the same but allowing '-' as well. It uses this to populate 'prefix'.

Autocomplete's suggestion is wrong for languages like PHP which have their own regex, is right for languages like HTML which should but don't, and is wrong for languages like Java which don't provide their own regex and which don't allow hyphens.

What we'll do is work around this mess right here as best we can: if the language-package provides its own regex which gives a different prefix from Autocomplete's, then we'll suggest that to the language service, and we'll patch the output of the language service to reflect this.

What Hack used to do in the past was provide its own THIRD regex, as an alternative to both the one provided by the Hack scope to Atom, and as an alternative to the one provided by Autocomplete. I could have continued with this (by augmenting LSP configuration with that third regex). But that puts more work on developers of language-providers than seems fair.

Note incidentally that VSCode has a word regex for autocomplete and the other editing features. Its regexes are all unicode-aware. Its default regex is more or less `[a-zA-Z0-9_]`. Its regex for Hack includes `$` and does not include `\`. I'm aiming to make the Nuclide environment as similar as possible to that of VSCode, so that language servers will work interchangeably between the two.

Reviewed By: hansonw

Differential Revision: D5374176

fbshipit-source-id: 8a4568dcb26eacc221bd848579cac73bd2a4fb85