AutocompleteCacher should filter+sort from the start
Summary:
Prior to this diff, if you had `foo.ba|` and then pressed ctrl+space to invoke autocomplete, then the underlying language service was responsible for providing the autocomplete suggestions and filtering and sorting them.

But as soon as you typed another character e.g. `foo.bar|`, then the `AutocompleteCacher.updateResults` function became responsible for filtering+sorting.

This seems unsatisfactory! Because it's a shame to have to put the filtering+sorting code in a common location where both server-side language service and client-side updateResults can call it. And the two pieces of code should be kept tightly in sync so that the user gets consistent sorting+filtering in both scenarios.

This diff achieves that by adding a new configuration option to AutocompleteCacherConfig. It is expected that languages will opt into both the new method `updateFirstResults` (which is called to filter+sort the results of `getSuggestions` the first time) and the eisting method `updateResults` (for subsequent times).

I'm not very happy with this API pattern. It would be better to factor them differently. But doing so would be pretty disruptive, so I'll leave that for Nat to do as part of his planned overhaul of AutocompleteCacher.

Reviewed By: nmote

Differential Revision: D5387036

fbshipit-source-id: fa319d05fa0c45910b2fc64e18a1600079a212e7