Add 'activatedAutomaticallyTrigger' to getAutocompleteSuggestions
Summary:
This diff makes an autocomplete request includes information about what key was just typed, so language servers can make an efficient decision about whether that key merits an automatic trigger of autocompletion.

Long story:

AutocompletePlus asks providers for as-you-type suggestions on every single keystroke so it can automatically display the suggestion list. Providers have the choice of either giving suggestions or not.

In the case of LSP, it has a notion of 'autocomplete triggerCharacters', with the idea that automatic display should only occur when (1) the user types one of the core alphanumeric characters that the editor always believes should automatically trigger autocomplete, or (2) the user types one of a small additional list of LSP-provided characters like '.' or '$' which should also automatically trigger autocomplete.

The way I'll implement this is by continuing to let AutocompletePlus ask for suggestions on every single character, but then LspLanguageService will consult its list of trigger characters and decline if necessary.

To do this, LspLanguageService will need to know what was the character just typed.

We don't know *directly* what was just typed that triggered an autocomplete request. The best we can do is figure out the character immediately to the left of the caret. I haven't been able to observe scenarios (e.g. pasting, deleting) where the two differ.

Which module should be the one that obtains that character?

It could be done on the nuclide-server side in response to an Autocomplete request. To do this it would have to obtain the current buffer -- which will only succeed if no intervening characters were typed and sent to nuclide-server after the moment the autocomplete request was generated. But I don't think that's reliable. Autocomplete is in a unique situation where: (1) we expect typing characters to come at us fast and furious; (2) an autocomplete *response* from several characters ago is still perfectly good and is in fact the normal expected state of affairs - the AutocompleteCacher remembers earlier responses from the language server and updates them to new states of the buffer; (3) it is normal and expected that autocomplete *requests* are too slow to come on every character.

(Note: we still at the moment require that no intervening characters were sent to the LSP server. That could be relaxed, by the same techniques as it's relaxed by the client-side autocomplete cacher, but doesn't seem worth it).

And so we come to this diff. It makes the client-side AutocompleteProvider capture the preceding character relating to the autocomplete request, and send it along with the 'isManualRequest' flag, so the langauge service can handle it without having to get nuclide-server-buffer-contents at the exact right fileversion.

Incidentally, there were surprisingly few places in the codebase to change. Only Flow, Hack and Java implement the LanguageService directly. All the other languages use SingleFileLanguageService, and I didn't bother plumbing the flag through to that, so they didn't need to be changed. (heck, even Java didn't need to be changed because it doesn't use the last two parameters, but I did so anyway for neatness).

Reviewed By: hansonw

Differential Revision: D5390866

fbshipit-source-id: 0af8e0a8a5a5e1827bbf2baa14ae8121bd2332b9