Fix unit tests that require mouse interaction with text editors
Summary:
Some specs interact with the text editor via mouse.

In 1.19 it seems that text editors don't have a height until the entire workspace is mounted; we need to borrow the integration test setup that mounts the entire workspace and sets a reasonable window size.

Note that there's an unusual bug with the Hyperclick spec where the spec hangs on teardown if we don't manually dismount the workspace. I haven't looked into it further but the workaround is easy enough ¯\_(ツ)_/¯

Reviewed By: matthewwithanm

Differential Revision: D5502576

fbshipit-source-id: 36c20e07b1cc574807737b76d45a3ec3f4ed992c