Show all app output in console pane
Summary:
Not great, but it's a start. This starts tailing the iOS Simulator log for the buck-ios plugin, and includes all output from com.apple.console, which is what FBLog statements use as Facility.

The tailing service has a shared config (annoying if you're only interested in React Native). We might need a better paradigm for long-running tasks like this (or maybe simply make them parametrizable?), but this is a quick way to add iOS logging.

Reviewed By: matthewwithanm

Differential Revision: D5510635

fbshipit-source-id: 07dcf3ee1405da043e94eb5fd77e7ccc3d61ddbd