Attempt to fix exceptions from stall tracking
Summary:
Due to our hack around delaying context menus, it appears that there might be some kind of race between showing a context menu as the window closes and the stall trackers.

Attempt to mitigate this by removing the `context-menu` listener as soon as the window close event is received. I'm not sure if this works, but it's worth a try (I can't repro)

Closes #1246

Reviewed By: matthewwithanm

Differential Revision: D5514466

fbshipit-source-id: 961ca5da11d22b67a1c05e4e503997876337e1fb