Add a small debounce to file change events for remote files
Summary:
I observed in one user's logs that they were getting 4 change events for every save, which causes an unnecessary amount of work on the client. I'm not really sure why this was happening, but it's easy to circumvent this with a small debounce.

Sidenote - Atom 1.19 adds this to the native text buffer by default. We won't be using htis class when we upgrade anyway, but this is a quick fix.

https://github.com/atom/text-buffer/blob/master/src/text-buffer.coffee#L132

Reviewed By: matthewwithanm

Differential Revision: D5528804

fbshipit-source-id: 8f98ca8cb3e737c04aaf92f8e4a12bc89a3b8cde