Fix highlight on horizontal scrolling in file-tree sections
Summary:
If the file-tree sidebar is not full wide enough to fit the file names and is scrolled horizontally the highlight width wouldn't match the width of the rows.

This diff fixes it. Admittedly, though, I have little idea what is it I'm doing, so feel free to commandeer and do it the right way if I'm way off.

Reviewed By: jgebhardt

Differential Revision: D5525129

fbshipit-source-id: 4b06167f1ca037576340c097f16bc831bbbea894