Introducing V1 of the Big Dig protocol.
Summary:
This is the first attempt to formalize the first version of the Big Dig protocol.
The principles are explained in `protocol-v1.md`, but the basic idea is that
Big Dig multiplexes multiple bi-directional endpoins over a single WebSocket
connection using a simple, human-readable header as the "envelope" for an
individual message.

This revision also starts to push us in a direction where those who build on
top of Big Dig do it via `BigDigClient` and `BigDigServer` so that we can start
to hide away some of the internal data structures. The goal is to help delineate
the public API of Big Dig.

Reviewed By: hansonw

Differential Revision: D5441624

fbshipit-source-id: 556992db973b5092870631a2e4ac28eea5260a43