Fix diff gutter for multiple working roots
Summary: Previously the logic to fetch the file diffs was dependent on the revisions cache to get the head revision id, and only the active repository data was being fetched. It is possible to fetch all revisions data for all the roots the user has open in Nuclide, but it would be very costly. Instead fetch only the needed head revision id.

Reviewed By: jgebhardt

Differential Revision: D5533118

fbshipit-source-id: b09ee93fcfcdd78c8bcf8575b1a2ede3d9598baa