Busy-signal service is a singleton; consumers shouldn't dispose it
Summary:
When folks consume the busy signal service, it looks like they were all disposing of it.

They shouldn't! when you consume an atom service, you only get a shared reference to a singleton instance of it!

This diff removes the callsites to the 'dispose' method. It retains the method itself as a no-op, for legacy compatibility.

Reviewed By: hansonw

Differential Revision: D5520747

fbshipit-source-id: d205366b24d6435e738ddaf4fdb598327996cfa6