HostService exposes busy-signal APIs
Summary:
The HostServices API is how Atom Language Services communicate with Atom. Here I'm adding APIs to use the busy service.

The design of this API was subtle! Here are the constraints:

(1) showProgress returns a Promise<Progress> rather than a Progress, because all RPC-able calls are async and therefore have to return either void or ConnectableObservable or Promise.

(2) I had tried to design it so that showActionRequired takes in a parameter 'titles: ConnectableObservable<string>' as the mechanism for allowing the caller to update to the progress message. But nuclide-rpc doesn't let you pass ConnectableObservable as argument.

(3) The showActionRequired doesn't use the same mechanism. That's because it needs to asynchronously notify the caller when its tooltip has been clicked. It feels like a return type of ConnectableObservable<void> is the best way to represent this.

We end up with two completely different mechanisms for indicating that the busy icon should be disposed. In 'showProgress' we do that by disposing the returned Progress object. In 'showActionRequired' we do that by unsubscribing from the returned ConnectableObservable. I couldn't think of any way to unify the two.

Reviewed By: hansonw

Differential Revision: D5457369

fbshipit-source-id: 168004b828a6a277e328ea0a1aae465882752a80