use busy indicator to show connection status
Summary:
Here are the states LSP might be in:

* "Starting..." while it's establishing the connection to the LSP server and awaiting a response to the Initialize request - in which case it should show a busy spinner in the status bar.

* "Running" when everything's running smoothly - no need to display anything.

* "Stopped" when something went wrong and we've given up trying to auto-restart it - in which case it should show an 'action required' icon in the status bar.

That's what this diff does. (There are a few other states that we treat similarly too: "Stopping..." and "StartFailed").

Note: I think there are numerous other places where we should show busy/action-required icons in the status bar. This diff is merely a start.

Reviewed By: hansonw

Differential Revision: D5466667

fbshipit-source-id: e8372f72f8a066fab13c70ebb7992d89515d0bd9