support LSP 'progress' and 'actionRequired'
Summary:
This diff adds support for two new [non-standard] LSP notifications sent from server to client: `window/progress` and `window/actionRequired`. The client responds to them by showing a busy-indicator or an action-required-indicator in the status bar.

If a connection to an LSP server gets closed, then any indicators for that LSP server instance will also be removed. I do this with a per-connection 'HostServices' object.

Incidentally, I've observed people with multiple "Hack is connecting..." or "Waiting for Hack server to be ready..." dialogs on screen. I think that was because one connection produced a dialog, then it crashed, and another connection was spun up, and then the second connection produced the same dialog. The work in this diff to have a per-connection HostServices object will ensure that we only get one such dialog onscreen at a time, as desired.

Reviewed By: arxanas

Differential Revision: D5516795

fbshipit-source-id: 60669f352752c510530c25cac4ef0ead9d0b8c60