Cleanup of progress-report when HostServices are disposed
Summary:
Inside HostServicesAggregator, for each observable that we return, we explicitly unsubscribe when 'childIsDisposed' gets fired.

This diff make HostServiesAggregator do the same for the only promise it returns: this promises should explicitly be voided once 'childIsDisposed' gets fired.

Actually, I made it return an "empty" progress object in this case. The other possibilities would be to return undefined (which is what Observable.toPromise does if the observable completes without ever having emitted a value) or to throw an exception. I figure that an empty progress report will be less error-prone.

This diff also fixes a potential double-dispose -- if someone gets a progress object out of their HostServices, and then disposes their HostServices followed by the progress object. It used to generate two disposals of the underlying progress object (wouldn't actually have had any bad effect) but now disposes it only once.

Reviewed By: hansonw

Differential Revision: D5546565

fbshipit-source-id: f16de1462395e3604ac2112af30b02429acf0955