Hide 'Open in diff view' shortcut in source control navigator
Summary: When clicking on a file from source control navigator, the diff view opens up. This click action already checks if the split view is already open and in those cases just updaes the diff view state rather than calling a command that would reload the diff view. This change removes the shortcut button to open in diff view for source control navigator. This effect is achieved as s side effect of honoring the option 'openInDiffViewOption' boolean that is passed to MultiRootFileTreeChanges, currently that boolean only controls if the open in diff view option is displayed in the context menu but does not actually hide the shortcuts.

Reviewed By: tjfryan

Differential Revision: D5558733

fbshipit-source-id: 2c17b642affced0c093614c232d19412b636b613