Rename rmdir to rimraf in fsPromise.
Summary:
Most of the methods in `fsPromise` are straight wrappers of the built-in `fs`
module. This was not the case for `rmdir()` because the `fs` version would pass
an `ENOENT` to the `callback`, but our `rmdir()` would not return a rejected
promise in that case because it used `rimraf` rather than `fs.rmdir` under the
hood. This renames our method to `rimraf` to reflect this exception.

We could provide a `fsPromise.rmdir()` that faithfully reflects `fs.rmdir()`,
but in practice, I suspect that would lead to user error more often than it
would help, so it's probably best to exclude it from the `fsPromise` API.

Reviewed By: hansonw

Differential Revision: D5582004

fbshipit-source-id: 7562b3e719ebd78fe7557da28ae42cdd3790288e