Trigger autocomplete suggestions when a period is typed
Summary:
D5374176 changed the prefix passed to getAutocompleteSuggestions to be the replacement prefix which will not contain the character '.'. Therefore, Flow autocomplete suggestions are no longer triggered by a period.

This diff gets the character from the buffer to restore Flow autocomplete to its previous behavior of offering suggestions when a '.' is typed.

Reviewed By: ljw1004

Differential Revision: D5587132

fbshipit-source-id: 0870bc767baadf11fc11f9a3893c62045409e578