Prevent FileTree rendering infinite loop
Summary:
In D5247771 I believe I introduced the possiblity of an infinte loop of FileTree rendering. It can be triggered by some combination of a file tree that gets in a state where it never registers itself as fully loaded (as warned about in `FileTreeStore._clearTrackedNodeIfNotLoading()`) and/or a file tree consisting only of collapsed project roots.

I haven't been able to exactly recreate this tate, but it looks like the loop that we see in the profile of the user who reported the issue, and reading through the code seems to indicate that it would be possible.

This check should shortcircut this loop.

The loop looks more or less like this:

```
FileTree.render
FileTree.componentDidUpdate
  FileTree._scrollToTrackedNodeIfNeeded
    FileTree.props.scrollToPosition
      (if tracked node still exists - AKA if still loading)
        FileTreeComponent._scrollToPosition
          FileTreeComponent.setState
            FileTreeComponent.render
              FileTree.render (LOOP!)
```

Reviewed By: hansonw

Differential Revision: D5602858

fbshipit-source-id: 5091a2e42e9ad9de4b3282e3cbcb1c1cdba79618