HostServicesAggregator disposal of promise, race
Summary:
Another bugfix for HostServicesAggregator.

This one is similar to the same thing we'd needed in LspLanguageService.js: https://github.com/facebook/nuclide/blob/master/pkg/nuclide-vscode-language-service-rpc/lib/LspLanguageService.js#L207

Scenario: the user has several projects on the same host and closes one of them. I didn't untangle exactly what happened next, but the long and short is that apparently the HostServices object was disposed first, and someone tried to set the busy signal status afterwards. Maybe they had asynchronous cleanup work to do. Anyway, they show a busy signal, and then get an exception when trying to add it to the already-disposed HostAggregate, and so the busy signal never gets disposed.

In practice I observed messages in the status bar saying "GraphQL is stopping" or "Hack is stopping" which never went away until I finally closed the entire nuclide-server.

This diff fixes that.

Reviewed By: hansonw

Differential Revision: D5582795

fbshipit-source-id: 322e4e1dec91ab727335f951ec53d4261bd1abfe