Diagnostics should start as soon as a file is opened
Summary:
The language service connections are currently lazy, so they won't start up until one of the Atom services explicitly gets a connection for something.

In the case of push-based diagnostics-only providers, this means that you'll never actually activate the service :(

At the very least, diagnostics should start being observed once a file of the matching grammar is opened. We can easily implement this by adding an editor observer on the side.

Reviewed By: ljw1004

Differential Revision: D5622102

fbshipit-source-id: dd06f5531ede503afc6e23210d96c87d81112249