use projectDir as CWD
Summary:
This diff makes LspLanguageService use the projectDir as the current working directory when spawning the LSP-server command.

This change was motivated by Hack, whose "hh_client" LSP-server-command is just a shell-script, and uses the current working directory to determine which real hh_client binary to launch, and then launches it.

But I feel justified in adding it to LspLanguageService because...

* The paradigm VSCode adopted is that the language's VSCode extension is invoked when VSCode wants to launch that language's server; the extension could look at the project directory, set current working directory, provide a callback for how LSP should launch its server, or do anything.
* The paradigm Nuclide adopted is that each language provides a smaller plain-old-data set of configuration parameters. Nuclide will use those parameters arbitrarily many times in the future, maybe on remote devservers, to launch the LSP server. There's a lot less flexibility here. So we have to make Nuclide able to address most of the needs of most different languages.

Reviewed By: hansonw

Differential Revision: D5622297

fbshipit-source-id: 1a271c3faffebaa040bc18d8ccbc2e7a52a90bab