Change nuclide-file-tree.revealFileOnSwitch to default to true.
Summary:
From the user research done by the Nuclide team, it appears that users are
having trouble accessing features in the context menu in the file tree. By
auto-focusing the entry for the active editor in the file tree, this should
make these easier to find.

Note that this behavior is the default in editors such as VS Code.

Changelog: Change the default value of the `nuclide-file-tree.revealFileOnSwitch` config to `true` so that, by default, the active file will be revealed in the file tree by default. This behavior can be disabled by unchecking "Reveal File on Switch" in Nuclide's settings.

Reviewed By: matthewwithanm

Differential Revision: D5622873

fbshipit-source-id: 806e0d17d080aa8ad34520fe50c105d17af0fdd3