Associate devices with the adb/sdb port they were found on (1/N)
Summary: To accomodate having multiple instances of adb/sdb running on multiple ports, we need to track which port a particular device came from. It is possible to have multiple devices attached to different adb instances with the same device name / serial. This change adds infra to have the device poller search an array of ports rather than just a single port, and tags each device with the port on which it was found before returning.  All of the routines in AdbService etc now take in a DeviceId which includes this port.

Reviewed By: a20012251

Differential Revision: D5515918

fbshipit-source-id: 455d4d20fe6e9b4ec95298f0c55dcc0c88d13fc8