Add a DeviceAction provider to Device Panel (7/N)
Summary: This change adds a provider that can register actions for a particular device in the device panel and updates the UI to show an action drop down next to a device in the device list so the user can click to trigger an action.  The action drop down only shows up for devices that have at least one action available. If no devices have any actions, the action column is not added to the table at all (currently, this will be the case for most devices).

Reviewed By: a20012251

Differential Revision: D5539000

fbshipit-source-id: 3b6da0efabdee849eeca74f6305cbd3faeae0e32