Code search: Add a wrapper for silversearcher
Summary:
This works now (I have to add error handling though, including the case when ag is not in the $PATH).

For the time being, I'd like some feedback on the usability of this or any detail I'm not aware of.

I actually want to add in the nuclide config a dropdown to select the grep like tool to use, like ag, ack or ripgrep.

Reviewed By: hansonw

Differential Revision: D5618393

fbshipit-source-id: 090ab1a6d78cca14929eb158e0e77c39ade7d29a