Fix race condition in Python servers
Summary:
Wow, this has gone unnoticed for a very long time! I was debugging the test flakiness and narrowed it down to this.

This is an instance of the classic set/get with awaits race. I moved the async-ness into the `JediServer` as we've done with other RPC processes.

I think this actually got worse after I bumped Atom to 1.19.1 because the text editor was improved, thus making the races more noticeable!

Reviewed By: matthewwithanm

Differential Revision: D5646615

fbshipit-source-id: 3ace739c19033a7a315d55f976bfcac0f2ddce99