Provide CodeActions for missing imports in the JS AutoImports LSP server.
Summary:
The fb-js-imports-server now provides CodeActions for missing imports. If requested for CodeActions in a range that is missing an import with a Diagnostic from this server, the server will now provide a code action with the command "add_import" and all required arguments to add the import.

Note that this diff does not contain code adding a handler to actually execute these commands so these code actions cannot actually be executed at the moment.

Reviewed By: hansonw

Differential Revision: D5509006

fbshipit-source-id: e8c6754eaa335a3d6298f4adcdef7f544edbaaf0