Stub out fetch in all tests
Summary:
Atom 1.19+ segfaults much more frequently during integration tests now.

I've discovered the root cause (https://github.com/electron/electron/issues/10051), but we probably won't get a fix for a while...
This function gets called when a test exits *just as* a web request comes back - so I think this is happening when we try to make web requests in integration tests.

I think it's fine if we just stub out all fetches in tests.
It's not safe to assume that fetches are going to work on CI, anyway.

Reviewed By: matthewwithanm

Differential Revision: D5656739

fbshipit-source-id: e80d8a954bb7d6b4b928f3de29707d5232452e3d