Reduce logspam from opening non-existent files
Summary:
Normally we try to watch a file when it's opened, but this logs an error if it doesn't exist yet.
This is a fairly normal circumstance so skip the logging (both in tests and in practice)

Note that I didn't edit the corresponding callsite for remote directories, since we shouldn't be trying to open non-existent directories.

Reviewed By: matthewwithanm

Differential Revision: D5660136

fbshipit-source-id: 5696b640896ceaadb4a96aa82056b9072a972e4b