Modernize remote project connection modal
Summary:
A bug in the management of these modals made it possible to hide the project connection dialog without properly cleaning it up.

Digging into the issue I found that this code was a little bitrotty; we've since developed patterns and utilities (like `bindObservableAsProps()` and `showModal()`) that avoid these problems and simplify logic drastically.

I tried not to go down the rabbit whole too deeply and didn't really touch the React components for either the connection dialog or create connection form. The main changes are:

- Scrap the complex management of multiple modals in favor of using a single one and swapping out its contents
- Centralize the state, Flux-style, but using a BehaviorSubject for simplicity

Reviewed By: hansonw

Differential Revision: D5656485

fbshipit-source-id: 7bf15dcdfa3dc694a2102f53a683e31ac8a321c7