Support windows properly
Summary:
I had to update the regex to support the paths that start with C:\
I had to put the options before the search string in the subprocess to be compliant with the windows format of rg
Only rg works on windows. Ag and ack have some issue on windows. It seems that the ports are somewhat old and the parameters that are send from nodejs are not parsed properly by the subprocesses. I spent a lot trying to figure out what to do. My conclusion is that these tools only work well from the command line

Reviewed By: wbinnssmith

Differential Revision: D5684604

fbshipit-source-id: 20b11c128bff30d580bc9930081c9a395378f183