Remove ObservableDiagnosticUpdater
Summary:
Observables are awesome but it's probably best to stick to callback +
disposable based APIs since they're super familiar to Atom users and
leak less API surface area (Rx operators). Plus they can be easily
converted to and from observables, which we can (and do!) do on either
side of the service boundary.

In a future diff I'll add reselect to the Selectors module so that we
do more intelligent caching.

Reviewed By: hansonw

Differential Revision: D5690063

fbshipit-source-id: 339166a44868b3d5fa09aaa41ebf4c1a7ce07a6c