Fix launching default android activity
Summary: Launching the default android activity for a package with no explicit activity name is not working (and the argument -N is invalid and causes an error). Switching to using monkey and adding the ability to set the "wait for debugger" bit via `adb shell am` prior to invoking the package.

Reviewed By: a20012251

Differential Revision: D5676340

fbshipit-source-id: e91ec95faf23363dd4e7cb82c36ecc9360132260