Fix device panel often shows "no devices connected" when refreshing
Summary: When I first switch to a connection, or if I click a device to view details and then click back to return to the device list, I frequently see the "no devices connected" message for a couple seconds before polling results come in.  This change introduces a pending state to an Expected which can indicate the Expected value has not been returned yet. The device panel will show a loading spinner while in this state.

Reviewed By: a20012251

Differential Revision: D5703545

fbshipit-source-id: 2f864dc15ee00abe92c0170870d8f5ec44419b52