Diagnostics: Add actions to fetch & set code actions for a diagnostic
Summary:
This adds two actions `fetchCodeActions` and `setCodeActions`, that manage requesting and setting the result respectively.

Note that until we have a real use case I've decided to assume that this we only care about the actions for one request at a time. this means:

Otherwise we'd have to worry about periodically emptying out the WeakMap.

Reviewed By: mostafaeweda

Differential Revision: D5712113

fbshipit-source-id: 925a694b62f4966c36765b6fab4ca4a95e628f3c