Ensure context menu actions are available when multiple files are selected
Summary:
Addressing T20733526 (File Tree: "Duplicate" option is not available when multiple files are selected)
and T20734638 (File Tree: 'Add to Mercurial' context menu item is not available when multiple files are selected)

Adding multiple files to mercurial work as you would expect. You select a few files and it will be added to mercurial.

For duplicating files, I decided to show a dialog for each file, so the user can name accordingly.

Reviewed By: captbaritone

Differential Revision: D5725600

fbshipit-source-id: 536553bd4b60056ab7418a134265892c85f47161